разбил модель экрана уведомлений на разные файлы
