+++ UIKitDZ/UIKitDZ/Notification/CommentCellType.swift
@@ -1,0 +1,14 @@
+// CommentCellType.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// комент простой
+struct CommentItemCell {
+    // аватар
+    var userImageName: String
+    // коментарий
+    var userText: String
+    // картинка поста
+    var imageName: String
+}
