+++ UIKitDZ/UIKitDZ/Notification/SubscribeCellType.swift
@@ -1,0 +1,12 @@
+// SubscribeCellType.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// комент с кнопкой
+struct SubscribeItemCell {
+    // картинка юзера
+    let imageUserName: String
+    // комментарий юзера
+    let userComment: String
+}
